Add a command line flag for activating simple mode.
The command line flag (-s) must be set to allow for the activation of
simple mode. This is required to prevent people accidently turning the
mode on simply by having mapped a switch to channel 6. This should
later be replaced by an option to activate/deactivate simple mode in
QGC.