qurt: added HEXAGON_TOOLS_ROOT variable
The environment variable HEXAGON_TOOLS_ROOT must be set to the location
of the installed HexagonTools. This could be a global install at
/opt/HEXAGON_Tools/7.2.10/Tools or it could be a personal install
${HOME}/Qualcomm/HEXAGON_Tools/7.2.10/Tools.

The default install path used by the installer is exported in the Vagrantfile.

There is still an issue that the hexagon linker fails on mmap of the shared folder.
If the shared folder was mounted via NFS for Linux and MacOS the problem would be
resolved.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>