px4iofirmware: ignore safety when override is enabled
this forces PWM outputs in override, ignoring the safety state. This
makes in-flight reboot safer. ArduPilot has had this behaviour for a
while, but it was done in the FMU code by overriding the safety
state. That was not ideal as it didn't work during bootup and was racy
at all times.