Y4 delta wing Tiltrotor Configuration
This is a quadrotor Y4 deltawing VTOL configuration as in 2 motors side
by side on the front and 2 motors coax on the rear. This also similar to
the firefly6 VTOL, except that front bottom two motors are absent and
the rear coaxial motors also have the provision to tilt apart from the
front ones.