navigator: don't set the MC cruise speed
Instead, let the multirotor position controller take care of respecting
the cruise velocity. If it is set in the navigator, there is the
drawback that the param value is only respected on a waypoint change
instead of immediately.

I have not looked into the fixedwing implementation and therefore left
it as is.