modules: mavlink: mission: Correctly handle state timeout
With 2 different system ids(QGC + custom application) handling mission
was causing getting mavlink_mission stuck forever.

The previous state timeout was just considering _time_last_recv and
not _time_last_sent also as it was the last "if else" and
_retry_timeout is smaller then _action_timeout it would never happen.

First thing was to move this if up but as _time_last_sent is updated
at each send function it would never reset the state when handling
MAVLINK_WPM_STATE_GETLIST. So another timeout variable was added to
just track and reset the state machine

Also I removed the calls to hrt_elapsed_time() and use the now value
saving one call to hrt_absolute_time().