Enable I2C2 bus of auav x21
Enable the I2C2 bus of auav x2.1. This change came after a MAG registering 3 times on the I2C bus due to messed initialization script on the x21 board.

We slightly moded the board so we could use the I2C2 bus but now there was no initialization of the bus.