drivers: Add driver for mb12xx_serial
mb12xx sonars also have the serial variation but it uses RS232 standard.
It do not work with regular UART modules unless the signal is inverted.

So here I'm introducing INVERT_UART_INPUT(uart, _invert_true) that
must be implemented by boards that wants to use this driver.

That is one of the reasons why this module was not integrated into
the I2C mb12xx the second is because it do not share any protocol,
the serial version don't need to request a registers is just receives
the distance.