Change the  compass calibration calculation
"result = px4_ioctl (fd, MAGIOCCALIBRATE, fd);" This sentence will cause the compass driver to self-calibrate.
The driver program of hmc5883 compass  can self-calibration. This will change the scale factor of hmc5883 driver. Therefore, compass calibration input data may have been multiplied by a scale factor.So we have to consider that the scale factor in the driver might not be 1.0.
If mscale.*_scale is 1.0f. Then it will not affect the calculation results. If mscale. * _ Scale is not 1.0, then take the scale of the driver program into account.