ekf2: Fix wind estimation bug.
This prevents a condition where depending on the method of entering into fixed wing flight it is possiblefor sideslip fusion to commence before airspeed fusion.
This can lead to bad wind estimates and rejection of airspeed measurements.
This patch ensures that in where airspeed sensing is being used, sideslip fusion will not be requested until wind estimation has started.