Adds the ability to add additional optional files to the ROMFS at compile time
When the ROMFS is built/compiled, the files listed in the romfs_optional_files
variable are copied into the ROMFS/px4fmu_common/extras directory within the build directory.

Once the firmware is uploaded, these ROMFS files are present in the etc/extras/ directory
on the Px4 board.

The romfs_optional_files variable is set in the board-specific cmake file.