added X-UAV Mini Talon V-Tail airframe and mixer
The X-UAV Mini Talon airframe with the
new type Plane V-Tail uses the new AAVVTWFF_vtail mixer.
The new mixer is derived from the AAVVTWFF mixer and is a
generic V-Tail mixer. I left the flaps and wheel although
there are no flaps and wheels on the mini talon to keep it generic.
The airframe configuration is derived from the Albatross A-Tail.
The mixer could not be reused because servo setup is different.
The Albatros could be used when all servos are reversed and the
V-Tail controls are flipped left<->right. The new mixer is
introduced in order to keep the servo output order identical.