px4_daemon: fix SITL for macOS
It turns out that poll is not implemented consistently across platforms,
especially when it comes to which bits are set for which event, for a
comparison, checkout:
https://www.greenend.org.uk/rjk/tech/poll.html

The takeaway as mentioned in the conlusion in the link above is that you
can't rely on POLLIN or POLLHUP to check if a pipe/socket has been
closed but you need to do a read and check for EOF.

Therefore, I removed the logic polling for POLLHUP which triggered to
early on macOS. POLLHUP was triggered on shutdown(SHUT_WR) of the
client side when really we were interested in the server side doing a
shutdown(SHUT_RDWR).

Instead, we can just do a shutdown and close in the server, once we're
done and have sent the return code to the client.