rcS tone_alarm: fix CBRK_BUZZER
tone_alarm was started before the parameters were loaded, and the first
tune was played before that as well. CBRK_BUZZER was then read as default,
ignoring the user-configured value.

We now start tone_alarm after we load the parameters. Note that a
previously published startup tone or SD card error will still be played.