simulator_mavlink: switch battery max and min to fix simultion
In fa3b60c99246219f58ecfdee9d0634fc6ba50c50 the simulated battery
percentage was reversed. I'm assuming because of the possibly
missleading variable name. Now I'm fixing it by switching the
maximum and minimum voltage such that the name is not misleading anymore
but it still works as expected.