FlightTasks: disable failsafe descend on RC loss
The problem with returning false and therefore causing a FlightTask
failsafe is that it does not follow the configuration that one might set
using the commander params.

Specifically if the RC failsafe behaviour is set to "Disabled" (0), then
the FlightTasks failsafe causes the drone to descend even though it
supposedly should not react at all.

With this change the behaviour is as follows with NAV_RCL_ACT = 0:
- POSCLT: stays in POSCTL but with all sticks 0, so it stays still.
- ALTCTL: stays in ALTCTL but with all sticks 0, so it stays horizontal
  and keeps altitude.
- MANUAL: stays in MANUAL but with all sticks 0, so it stays horizontal
  but throttle will be half-way which either leads to a descent or fly
  away. This case is clearly not optimal.