Allow to keep mag calibration after airframe change
The magnetometer calibration is erased when the airframe config is updated. This is based on the assumption that a different airframes have different magnetic influence.

This commit adds the parameter CBRK_KEEP_CALMAG to preserve the magnetometer calibration on airframe change.
Use it if you do not change the physical airframe (only the software airframe config), and do not want to recalibrate.