Deletion of <<continue>> in while loop and do calculations whether we have or not a POLLIN event and detection of tasks which missed their deadlines.
In certain tasks, we don't jump to the next iteration if the result of px4_poll is not a POLLIN event as it used to be, instead we go through the loop and do calculations as we had a POLLIN event.
In every task activation function, check whether the task has missed its deadline or not plus the number of deadline missed.