Remove fmu-v2_default from build
Due to a silicon errata on STM32F427VIT6 we can only use 1 MB flash
instead of 2 MBs that would be available on the chip.

The target fmu-v2 was used for Pixhawks with the silicon errata while
the target fmu-v3 can be used for newer Pixhawks without the errata.

For several years we have maintained the fmu-v2_default target by
optimizing flash space and removing debug helpers as well as drivers for
less common peripherals and sensors.

Now that the release v1.10 is forked off, it's sensible to sunset the
fmu-v2_default target in order to make development easier.

This only removes the _default target which contains multicopter,
fixedwing, and VTOL modules but keeps the separate _multirotor,
_fixedwing, _rover modules.