Gimbal: Reversing the pitch output for pixy f
After sending MAV_CMD_DO_SET_ROI_LOCATION to the drone, the gimbal
looked up into the sky instead of looking down on the Region of
Interest. In a ground test, the gimbal looked even further up if the
altitude of the ROI was lowered. Looks like the Pixy F gimbal
understands pitch reversely compared to px4.