mavlink: lock sending messages outside
This is an attempt to prevent any of the 3 parts of a MAVLink message to
end up on the buffer without the other parts. When partial messages end
up on the buffer it will cause message drops because the parser reads
too far and will miss the subsequent message.

By locking around the full process and preventing races between the
message senders in the receiver thread and the actual sending threads we
prevent message drops which e.g. leads to a cleaner initial param
download.