cuav/nora: Enable UAVCAN
As done for the hex/cube-orange. As far as I understood the limitation in the past was due to the MCU family. As far as I understand from some conversations here, it has been solved it therefore it should work here, too.

Not tested on real hardware. Just doing some research here.