Handle actuator outputs to have 8 motors by default
Previously, if a vehicle type was unknown, it would default to the fixedwing actuator output configuration.

However, this doesn't seem very intuitive, since fixedwings should be the exception of defining the actuator outputs.

This commit reverses the relation ship of the if else and utilizes the switch case to be proper