px4io: circuit breaker set -> safety has no influence
The IO firmware does not support completely disabling the IO safety.
#1399 and
8ea5fd20c1c268c6baf145606186592bf5bd3699 introduced just setting
safety off once when the circuit breaker gets set and on boot but
the button is still active and can turn safety on again
and even disarm the vehicle.

I'm adapting the implementation of the SafetyButton driver on the FMU
that was introduced in
35fc51de920c6ca01806a044d3d37b01f8a5758a