BNP055 driver - first working version
This is a driver for the Bosch BNO055 IMU, which is connected
via I2C. This chip actually contains a MCU to calibrate itself
and run a sensor fusion algorithm, so it can output complete
orientation values. However, this driver does not use this
feature, since PX4 already has a (probably better) module for
that.

This means that for this driver, the IMU is just an
accelerometer, a gyroscope and a magnetometer that happen to
be on the same module. The driver makes use of the integrated,
configurable low-pass filters on the sensor and outputs at
250 Hz (accel, gyro) and 30 Hz (mag).

This PX4 driver is based on the original Bosch C driver.