Moved Console from UART7 to UART8
The objective was to disable the Console (as UART7 is needed due to USART6 not available/working) but by omitting the Console line completely it stops the Control Zero from booting. Moving this to UART8 (which is Frsky Telemetry and likely used sparsely makes the most sense).

Tested and works.