cdev: Fix deallocating devname in nuttx protected/kernel builds
The CDevs always live in kernel side. In kernel the devname
is allocated in kernel heap. Replace "free" with kmm_free on nuttx to
correct the deallocation; this automatically maps to "free" on flat build.

Define __KERNEL__ flag so that the proper nuttx headers are used for building.

TODO: It is fishy in the first place that the devname is allocated in inherited
classes but deleted in the CDev base class. Especially when the devname is
often passed as a constant string (which cannot even be freed) by many drivers.

The allocation and deallocation of devname should be done within the CDev base class
instead.

Signed-off-by: Jukka Laitinen <jukkax@ssrc.tii.ae>