Benewake TFmini drivers
I have modified the drivers of TFmini because of two reasons. One is because TFmini was the earliest product with a range of just 12m. It's manufacturing is already stopped. Now Benewake has several LiDARs with different ranges the maximum is 350m with improved blind zone. I modified the driver range section according to the latest LiDARs. Second reason is the new LiDARs also contain byte information for chip temperature. There are no more reserved bytes as it was in case of mini. So I modified those bytes as well. I have already tested it with different LiDARs.