SSRC: Store the px4 parameters which are specific to flying location in git
There are some specific settings needed to get safe and precise flying in small
indoors location. These parameters can be stored on a config.txt file on the drone
sdcard; this is just to store those parameters also in the git for tracking.

NOTE: this is not the best location for such settings, and definitely not for
upstreaming or such

Signed-off-by: Jukka Laitinen <jukkax@ssrc.tii.ae>