Allow overriding sitl_gazebo build_cores with environment variable BUILD_CORES
This is far from a perfect solution, it simply tries to offer a minimum of flexibility when deploying automated builds.

This allows the possibility of overriding the parallelism when building sitl_gazebo, which is memory intensive and can result in oom ( see https://github.com/PX4/PX4-Autopilot/issues/17691 ). This is especially the case as new CPUs have more and more logical cores, but systems still generally have 16GB or less of RAM. In my 16GB system, `master` currently uses 14 parallel jobs which results in oom and with no way of resolving it since it is hardcoded to N-2 jobs in `ExternalProject_Add`.

Example of usage: `BUILD_CORES=8 DONT_RUN=1 make px4_sitl_default gazebo`

Important caveat with this solution. Unless the CMake file is reloaded, only the first time the example above is run will affect the number of parallel jobs. Trying to set a different `BUILD_CORES` environment variable afterwards does not work, because `ExternalProject_Add` is already defined. I was unable to find a workaround or a better solution.