Commander: fix the case where the user does an RC override while in data link failsafe
which is most of the time a return mode switching the main state.
Without handling the failsafe also in return mode the faisafe state is
cleared and it's not distinguishable from a normal RTL.

So in RTL there's also data link loss handling now and RC override is
not allowed during data link loss failsafe.