ekf2: remove range aid functionality
Range aid was introduced to stabilize the EKF estimates during takeoff
and landind of large vehicle having their barometer highly affected by
ground effects.
This feature has undesired effects on the reference altitude given that
it produces measurements relative to a different datum
(i.e.: the ground). While switching in and out of range aid, the estimator
accumulates an offset that can grow over time (known as ratcheting),
leading to an undesired offset of the estimated global position.

After discussion, it has been agreed that the range finder should only
be used as an assistance for takeoff and landing in the height
controller and not as an aiding source for navigation.