batmon: Adapt after smbus_sbs changes
With the refactor of smbus_sbs, most functionality has been moved to the
base class. Remaining in this driver is getting the cell count, reading
out the individual cells, and calculating the max delta between cells.

This commit removes suspend and resume functionality for now assuming
that's not actually used.

Signed-off-by: Alex Mikhalev <alex@corvus-robotics.com>