Signed-off-by: Jorge Zehl <jorgeignacio.lozanozehl@nxp.com> mc_pos_control: Create algorithm for speed limiting
What has been added is a practical lowspeedmode__state_update function that consists of a state machine
that changes state every time when the low speed siwtch is activated by the user through the rc transmitter.
The state machine updates the values of the parameters MPC_XY_VEL_CRUISE, MPC_XY_VEL_ALL, MPC_XY_VEL_MANUAL, MPC_XY_VEL_MAX.
The new values are replaced by the value of a new parameter MPC_XY_VEL_SLO which has a standard value of 3 m/s.

When the swtich goes from active to inactive position or the drone is turned off the previous values will be rewritten,
for this same reason 4 new parameters MPC_CRUISE_OLD, MPC_VEL_ALL_OLD, MPC_MAX_VEL_OLD, MPC_MANUAL_OLD were created which are used to store the values in the non-volatile memory.

The idea of creating this is the regulations in Europe to limit the speed in case of people in the surroundings.
For more information see page 26 of the following link https://eur-lex.europa.eu/legal-content/EN/TXT/PDF/?uri=CELEX:32019R0945&from=EN.

The good thing about this change is that it allows the user to fly according to the rules. The user can see if the limit is activated by the MPC_LOWVEL_STATE parameter.

The changes are the following:
- New function lowspeedmode__state_update(): State machine that is the brain to know if the limit will be activated or not.
- New subscription to the topical action_request to know the action that the user requires ACTION_LOW_SPEED_MODE_ON or ACTION_LOW_SPEED_MODE_OFF
- 4 new parameters MPC_CRUISE_OLD, MPC_MANUAL_OLD, MPC_VEL_ALL_OLD, MPC_MAX_VEL_OLD that are used as non-volatile memory
- New parameter MPC_XY_VEL_SLO the speed that the drone will maintain when the switch is on
- New parameter MPC_LOWVEL_STATE to provide the possibility to the user to visualize if the low speed mode is activated or not